fix: bug-fix for payment part

+++ usermicroservice/src/main/java/com/hacettepe/usermicroservice/dto/PaymentInfoDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class PaymentInfoDTO {
-    private int cardNumber;
+    private long cardNumber;
     private int cvc;
     private int expirationMonth;
     private int expirationYear;
