+++ usermicroservice/src/main/java/com/hacettepe/usermicroservice/repository/IPaymentInfoRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 public interface IPaymentInfoRepository extends JpaRepository<PaymentInfo, Long> {
     @Query(value = "UPDATE payment_infos SET stripe_token = :stripeTokens WHERE card_number = :cardNumber", nativeQuery = true)
-    void addStripeToken(int cardNumber, String stripeToken);
+    void addStripeToken(long cardNumber, String stripeToken);
 
-    PaymentInfo findPaymentInfoByCardNumber(int cardNumber);
+    PaymentInfo findPaymentInfoByCardNumber(long cardNumber);
 }
